requirements.txt
Babel==2.9.1
chardet==3.0.4
decorator==4.4.2
docutils==0.16
ebaysdk==2.1.5
freezegun==0.3.15
gevent==22.10.2
greenlet==2.0.2
html2text==2020.1.16
idna==2.10
Jinja2==3.1.2
libsass==0.22.0
lxml==4.9.3
MarkupSafe==2.1.3
num2words==0.5.12
ofxparse==0.20
passlib==1.7.4
Pillow==10.2.0
polib==1.1.1
# psutil==5.8.0  # Temporarily commented out
psycopg2==2.9.9
pydot==1.4.2
python-dateutil==2.8.2
python-stdnum==1.19
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
urllib3==2.0.3
vobject==0.9.6.1
Werkzeug==2.3.6
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
zeep==4.2.1
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2 